Simple Java Math Example
This is a small project understanding Classes and Methods in Java. I
have created two different classes one is the main class and the other
is a class that holds all the math algorithm. The main class extends the
algorithm class to be able to call all of it's methods.